requirements.txt
mysql-connector-python==8.0.16
numpy==1.16.4
pandas==0.24.2
requests==2.22.0
xlrd==1.2.0
